+++ .prettierrc
@@ -1,7 +1,9 @@
 {
-  "tabWidth": 2,
-  "semi": true,
-  "singleQuote": true,
-  "printWidth": 100,
-  "jsxSingleQuote": false
+    "trailingComma": "es5",
+    "tabWidth": 4,
+    "semi": true,
+    "singleQuote": true,
+    "printWidth": 120,
+    "jsxSingleQuote": false,
+    "bracketSpacing": true
 }
