+++ src/basic/components/CartItem.js
@@ -1,9 +1,9 @@
 import { createElement } from '../utils/createElement';
 
 const CartItem = () => {
-  return createElement('div', {
-    id: 'cart-items',
-  });
+    return createElement('div', {
+        id: 'cart-items',
+    });
 };
 
 export default CartItem;
