+++ src/basic/components/CartTotal.js
@@ -1,10 +1,10 @@
 import { createElement } from '../utils/createElement';
 
 const CartTotal = () => {
-  return createElement('div', {
-    id: 'cart-total',
-    className: 'text-xl font-bold my-4',
-  });
+    return createElement('div', {
+        id: 'cart-total',
+        className: 'text-xl font-bold my-4',
+    });
 };
 
 export default CartTotal;
