+++ src/basic/utils/createElement.js
@@ -1,15 +1,15 @@
 export const createElement = (tag, { id, className, textContent }) => {
-  const $element = document.createElement(tag);
+    const $element = document.createElement(tag);
 
-  if (id) {
-    $element.id = id;
-  }
-  if (className) {
-    $element.className = className;
-  }
-  if (textContent) {
-    $element.textContent = textContent;
-  }
+    if (id) {
+        $element.id = id;
+    }
+    if (className) {
+        $element.className = className;
+    }
+    if (textContent) {
+        $element.textContent = textContent;
+    }
 
-  return $element;
+    return $element;
 };
